+++ packages/apps/stock/src/screens/supplierArrivals/SupplierArrivalSelectTrackingScreen.js
@@ -16,74 +16,18 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, {useCallback, useState} from 'react';
-import {View, StyleSheet} from 'react-native';
-import {
-  Card,
-  HeaderContainer,
-  Icon,
-  PopUpOneButton,
-  Screen,
-  Text,
-  useThemeColor,
-} from '@axelor/aos-mobile-ui';
-import {
-  ScannerAutocompleteSearch,
-  useDispatch,
-  useSelector,
-  useTranslator,
-} from '@axelor/aos-mobile-core';
-import {StockMoveHeader} from '../../components';
-import {filterTrackingNumber} from '../../features/trackingNumberSlice';
-import {displayItemTrackingNumber} from '../../utils/displayers';
+import React, {useState} from 'react';
+import {HeaderContainer, PopUpOneButton, Screen} from '@axelor/aos-mobile-ui';
+import {useTranslator} from '@axelor/aos-mobile-core';
+import {StockMoveHeader, SupplierArrivalAddTracking} from '../../components';
 import StockMove from '../../types/stock-move';
 
-const trackingScanKey = 'tracking_supplier-arrival-select';
-
 const SupplierArrivalSelectTrackingScreen = ({route, navigation}) => {
-  const Colors = useThemeColor();
   const I18n = useTranslator();
   const supplierArrival = route.params.supplierArrival;
   const supplierArrivalLine = route.params.supplierArrivalLine;
   const product = route.params.product;
-  const {trackingNumberList} = useSelector(state => state.trackingNumber);
   const [isVisible, setVisible] = useState(false);
-  const dispatch = useDispatch();
-
-  const fetchTrackingAPI = useCallback(
-    filter => {
-      dispatch(
-        filterTrackingNumber({productId: product.id, searchValue: filter}),
-      );
-    },
-    [dispatch, product.id],
-  );
-
-  const handleTrackingNumberSelection = item => {
-    if (item !== null) {
-      if (
-        supplierArrivalLine != null &&
-        item.id !== supplierArrivalLine.trackingNumber?.id
-      ) {
-        setVisible(true);
-      } else {
-        navigation.navigate('SupplierArrivalLineDetailScreen', {
-          supplierArrivalLine: supplierArrivalLine,
-          supplierArrival: supplierArrival,
-          product: product,
-          trackingNumber: item,
-        });
-      }
-    }
-  };
-
-  const handleAddTrackingNumber = () => {
-    navigation.navigate('SupplierArrivalAddTrackingScreen', {
-      supplierArrivalLine: supplierArrivalLine,
-      supplierArrival: supplierArrival,
-      product: product,
-    });
-  };
 
   return (
     <Screen removeSpaceOnTop={true}>
@@ -104,34 +48,13 @@ const SupplierArrivalSelectTrackingScreen = ({route, navigation}) => {
           />
         }
       />
-      <View style={styles.stockView}>
-        <Card style={styles.cardProductInfo}>
-          <Text>{product?.name}</Text>
-        </Card>
-        <View style={styles.trackingNumber}>
-          <Text style={styles.text_secondary}>
-            {I18n.t('Stock_AddTrackingNumber')}
-          </Text>
-          <Icon
-            name="plus"
-            color={Colors.primaryColor.background}
-            size={24}
-            style={styles.action}
-            touchable={true}
-            onPress={handleAddTrackingNumber}
-          />
-        </View>
-        <ScannerAutocompleteSearch
-          objectList={trackingNumberList}
-          onChangeValue={item => handleTrackingNumberSelection(item)}
-          fetchData={fetchTrackingAPI}
-          displayValue={displayItemTrackingNumber}
-          scanKeySearch={trackingScanKey}
-          placeholder={I18n.t('Stock_TrackingNumber')}
-          isFocus={true}
-          changeScreenAfter={true}
-        />
-      </View>
+      <SupplierArrivalAddTracking
+        navigation={navigation}
+        product={product}
+        setVisible={setVisible}
+        supplierArrival={supplierArrival}
+        supplierArrivalLine={supplierArrivalLine}
+      />
       <PopUpOneButton
         visible={isVisible}
         title={I18n.t('Auth_Warning')}
@@ -142,28 +65,5 @@ const SupplierArrivalSelectTrackingScreen = ({route, navigation}) => {
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  cardProductInfo: {
-    marginVertical: '2%',
-    marginHorizontal: 16,
-  },
-  trackingNumber: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    marginRight: 16,
-    marginBottom: 5,
-  },
-  stockView: {
-    marginTop: '2%',
-  },
-  text_secondary: {
-    fontSize: 14,
-  },
-  action: {
-    marginLeft: 10,
-  },
-});
 
 export default SupplierArrivalSelectTrackingScreen;
